+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.10.1-preview] - 2021-05-14
+
+### Changed
+
+- Auto itemization of constants and variables is no longer on by default. Use the respectively the `BoolPref.AutoItemizeConstants` and `BoolPref.AutoItemizeVariables` preferences to control the behavior. 
+
+### Fixed
+
+- Bug that needlessly triggered full UI rebuild in certain conditions.
+
+
 ## [0.10.0-preview] - 2021-04-30
 
 ### Added